Refactor SqliteExporterTests and SqlitePoiRepositoryGetByAmenityTests to use SqliteTestDatabase.CreateAsync for database initialization
