+++ tests/POIneer.Render.IntegrationTests/Infrastructure/Sqlite/PoiRepository/SqlitePoiRepositoryGetByAmenityTests.cs
@@ -1,4 +1,6 @@
+using POIneer.Render.Infrastructure.Sqlite;
 using POIneer.Render.TestHelpers;
+using POIneer.Render.TestHelpers.Sqlite;
 using Xunit;
 
 namespace POIneer.Render.IntegrationTests.Infrastructure.Sqlite.PoiRepository;
@@ -9,8 +11,8 @@ public sealed class SqlitePoiRepositoryGetByAmenityTests
     public async Task GetByAmenityAsync_ReturnsMatchingPois()
     {
         // Arrange
-        await using var tempDir = TestTemporaryDirectories.Create("GetByAmenityAsync_ReturnsMatchingPois", false);
-
+        await using var tempDir = TestTemporaryDirectories.Create("GetByAmenityAsync_ReturnsMatchingPois", true);
+        var dbPath = await SqliteTestDatabase.CreateAsync(tempDir);
 
         // Act
