+++ tests/POIneer.Render.TestHelpers/Sqlite/TestDatabase.cs
@@ -1,4 +1,8 @@
 using Microsoft.Data.Sqlite;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using POIneer.Render.Infrastructure.FileSystem;
+using POIneer.Render.Infrastructure.Flyway;
 
 namespace POIneer.Render.TestHelpers.Sqlite;
 
@@ -12,5 +16,40 @@ public static class SqliteTestDatabase
             // Important for tests to ensure that each connection is independent and doesn't interfere with others
             Pooling = false
         }.ToString();
+    }
+
+    public static async Task<string> CreateAsync(
+        TemporaryDirectory tempDir,
+        string fileName = "poi.sqlite",
+        CancellationToken ct = default)
+    {
+        var dbPath = Path.Combine(tempDir.DirectoryPath, fileName);
+
+        var options = Options.Create(new FlywayOptions
+        {
+            Executable = "flyway",
+            ConfigFileRelativePath = "migrations/flyway-poi.toml",
+            Debug = false
+        });
+
+        var processRunner = new ProcessRunner();
+
+        var invocationBuilder =
+            new FlywayInvocationBuilder(options);
+
+        using var loggerFactory = LoggerFactory.Create(builder => { });
+
+        var logger =
+            loggerFactory.CreateLogger<FlywaySqliteDatabaseInitializer>();
+
+        var initializer =
+            new FlywaySqliteDatabaseInitializer(
+                processRunner,
+                invocationBuilder,
+                logger);
+
+        await initializer.InitializeAsync(dbPath, ct);
+
+        return dbPath;
     }
 }
